requirements.txt
Django==3.1.5
djangorestframework==3.12.2
pyjwt==2.0.1
django-mptt==0.11.0
